refactor: Reestruturação das implementações e interfaces.
