+++ dbpautas-back-end/src/main/java/com/db/dbpautasbackend/controller/PautaController.java
@@ -7,8 +7,8 @@ import com.db.dbpautasbackend.enums.Categoria;
 import com.db.dbpautasbackend.enums.Voto;
 import com.db.dbpautasbackend.mapper.PautaMapper;
 import com.db.dbpautasbackend.model.Pauta;
-import com.db.dbpautasbackend.service.interfaces.ContabilizacaoService;
-import com.db.dbpautasbackend.service.interfaces.PautaService;
+import com.db.dbpautasbackend.service.ContabilizacaoService;
+import com.db.dbpautasbackend.service.PautaService;
 import io.swagger.v3.oas.annotations.Operation;
 import jakarta.validation.Valid;
 import lombok.AllArgsConstructor;
