+++ dbpautas-back-end/src/main/java/com/db/dbpautasbackend/repository/UsuarioRepository.java
@@ -1,7 +1,7 @@
 package com.db.dbpautasbackend.repository;
 
 import com.db.dbpautasbackend.model.Usuario;
-import com.db.dbpautasbackend.info.interfaces.UserDetailsInfo;
+import com.db.dbpautasbackend.info.UserDetailsInfo;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.stereotype.Repository;
@@ -11,7 +11,7 @@ import java.util.Optional;
 @Repository
 public interface UsuarioRepository extends JpaRepository<Usuario, Long> {
 
-    @Query("SELECT NEW com.db.dbpautasbackend.info.UserDetailsInfoImpl(u.cpf, u.senha, u.papel) FROM Usuario u WHERE u.cpf = :cpf")
+    @Query("SELECT NEW com.db.dbpautasbackend.info.impl.UserDetailsInfoImpl(u.cpf, u.senha, u.papel) FROM Usuario u WHERE u.cpf = :cpf")
     Optional<UserDetailsInfo> findUserDetailsByCpf(String cpf);
 
     Optional<Usuario> findByCpf(String cpf);
