+++ dbpautas-back-end/src/main/java/com/db/dbpautasbackend/service/ContabilizacaoService.java
@@ -1,4 +1,4 @@
-package com.db.dbpautasbackend.service.interfaces;
+package com.db.dbpautasbackend.service;
 
 import com.db.dbpautasbackend.enums.Decisao;
 import com.db.dbpautasbackend.model.Pauta;
