+++ dbpautas-back-end/src/main/java/com/db/dbpautasbackend/service/impl/ContabilizacaoServiceImpl.java
@@ -1,8 +1,8 @@
-package com.db.dbpautasbackend.service;
+package com.db.dbpautasbackend.service.impl;
 
 import com.db.dbpautasbackend.enums.Decisao;
 import com.db.dbpautasbackend.model.Pauta;
-import com.db.dbpautasbackend.service.interfaces.ContabilizacaoService;
+import com.db.dbpautasbackend.service.ContabilizacaoService;
 import org.springframework.stereotype.Service;
 
 @Service
