+++ dbpautas-back-end/src/main/java/com/db/dbpautasbackend/service/impl/UserDetailsServiceImpl.java
@@ -1,7 +1,7 @@
-package com.db.dbpautasbackend.service;
+package com.db.dbpautasbackend.service.impl;
 
 import com.db.dbpautasbackend.repository.UsuarioRepository;
-import com.db.dbpautasbackend.info.interfaces.UserDetailsInfo;
+import com.db.dbpautasbackend.info.UserDetailsInfo;
 import lombok.AllArgsConstructor;
 import org.springframework.security.core.GrantedAuthority;
 import org.springframework.security.core.authority.SimpleGrantedAuthority;
