+++ dbpautas-back-end/src/main/java/com/db/dbpautasbackend/service/impl/UsuarioServiceImpl.java
@@ -1,8 +1,8 @@
-package com.db.dbpautasbackend.service;
+package com.db.dbpautasbackend.service.impl;
 
 import com.db.dbpautasbackend.model.Usuario;
 import com.db.dbpautasbackend.repository.UsuarioRepository;
-import com.db.dbpautasbackend.service.interfaces.UsuarioService;
+import com.db.dbpautasbackend.service.UsuarioService;
 import lombok.AllArgsConstructor;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
