+++ dbpautas-back-end/src/test/java/com/db/dbpautasbackend/controller/LoginControllerIntegrationTest.java
@@ -4,7 +4,7 @@ import com.db.dbpautasbackend.dto.LoginDTO;
 import com.db.dbpautasbackend.dto.LoginRespostaDTO;
 import com.db.dbpautasbackend.enums.Papel;
 import com.db.dbpautasbackend.fixture.LoginDTOFixture;
-import com.db.dbpautasbackend.service.interfaces.TokenService;
+import com.db.dbpautasbackend.service.TokenService;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
