+++ dbpautas-back-end/src/test/java/com/db/dbpautasbackend/service/UserDetailsServiceImplTest.java
@@ -1,8 +1,9 @@
 package com.db.dbpautasbackend.service;
 
 import com.db.dbpautasbackend.fixture.UserDetailsInfoFixture;
-import com.db.dbpautasbackend.info.UserDetailsInfoImpl;
+import com.db.dbpautasbackend.info.impl.UserDetailsInfoImpl;
 import com.db.dbpautasbackend.repository.UsuarioRepository;
+import com.db.dbpautasbackend.service.impl.UserDetailsServiceImpl;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
